+++ cloud/acp-admin-cloud-constant/src/main/kotlin/pers/acp/admin/constant/RouteConstant.kt
@@ -5,13 +5,13 @@ package pers.acp.admin.constant
  * @since JDK 11
  */
 object RouteConstant {
-    const val UPDATE_ROUTE_INPUT = "update_route_input"
-    const val UPDATE_GATEWAY_OUTPUT = "update_route_output"
+    const val UPDATE_ROUTE_INPUT = "updateRouteConsumer"
+    const val UPDATE_GATEWAY_OUTPUT = "updateRouteProducer"
     const val UPDATE_ROUTE_DESCRIPTION = "update-gateway-routes"
     const val ROUTES_DEFINITION_KEY = "acp_cloud_gateway_routes_definition_key"
     const val UPDATE_GATEWAY_ROUTES = "update_gateway_routes"
-    const val ROUTE_LOG_INPUT = "route_log_input"
-    const val ROUTE_LOG_OUTPUT = "route_log_output"
+    const val ROUTE_LOG_INPUT = "routeLogConsumer"
+    const val ROUTE_LOG_OUTPUT = "routeLogProducer"
     const val ROUTE_LOG_DESCRIPTION = "route-log"
     const val ROUTE_LOG_CONSUMER_GROUP = "route-log-group"
 }