+++ cloud/acp-admin-cloud-dependencies/src/main/kotlin/pers/acp/admin/common/base/BaseFeignHystrix.kt
@@ -1,6 +1,6 @@
 package pers.acp.admin.common.base
 
-import feign.hystrix.FallbackFactory
+import org.springframework.cloud.openfeign.FallbackFactory
 import pers.acp.spring.boot.interfaces.LogAdapter
 
 /**