+++ cloud/acp-admin-cloud-dependencies/src/main/kotlin/pers/acp/admin/common/event/NoticeBusEvent.kt
@@ -6,4 +6,12 @@ import org.springframework.cloud.bus.event.RemoteApplicationEvent
  * @author zhang by 19/03/2019
  * @since JDK 11
  */
-class NoticeBusEvent(origin: String?, destination: String?, val message: String, source: Any = Object()) : RemoteApplicationEvent(source, origin, destination)
+class NoticeBusEvent(
+    originService: String?,
+    destinationService: String?,
+    val message: String,
+    source: Any = Object()
+) : RemoteApplicationEvent(
+    source, originService,
+    DEFAULT_DESTINATION_FACTORY.getDestination(destinationService)
+)