+++ cloud/log-server/src/main/kotlin/pers/acp/admin/log/conf/LogConfiguration.kt
@@ -3,16 +3,15 @@ package pers.acp.admin.log.conf
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.AutoConfigureBefore
-import org.springframework.cloud.stream.annotation.EnableBinding
 import org.springframework.cloud.stream.config.BindingProperties
 import org.springframework.cloud.stream.config.BindingServiceConfiguration
 import org.springframework.cloud.stream.config.BindingServiceProperties
+import org.springframework.cloud.stream.function.StreamFunctionProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.util.MimeTypeUtils
 import pers.acp.admin.constant.RouteConstant
-import pers.acp.admin.log.consumer.RouteLogInput
-import pers.acp.admin.log.consumer.instance.RouteLogConsumer
+import pers.acp.admin.log.consumer.RouteLogConsumer
 import pers.acp.admin.log.domain.LogDomain
 import pers.acp.spring.boot.interfaces.LogAdapter
 import javax.annotation.PostConstruct
@@ -23,9 +22,12 @@ import javax.annotation.PostConstruct
  */
 @Configuration(proxyBeanMethods = false)
 @AutoConfigureBefore(BindingServiceConfiguration::class)
-@EnableBinding(RouteLogInput::class)
 class LogConfiguration @Autowired
-constructor(private val bindings: BindingServiceProperties) {
+constructor(
+    private val bindingServiceProperties: BindingServiceProperties,
+    private val streamFunctionProperties: StreamFunctionProperties
+) {
+    private val routeLogConsumerBindName = "${RouteConstant.ROUTE_LOG_INPUT}-in-0"
 
     @PostConstruct
     fun init() {
@@ -33,11 +35,16 @@ constructor(private val bindings: BindingServiceProperties) {
     }
 
     private fun initConsumer() {
-        if (this.bindings.bindings[RouteConstant.ROUTE_LOG_INPUT] == null) {
-            this.bindings.bindings[RouteConstant.ROUTE_LOG_INPUT] = BindingProperties()
+        if (this.bindingServiceProperties.bindings[routeLogConsumerBindName] == null) {
+            this.bindingServiceProperties.bindings[routeLogConsumerBindName] = BindingProperties()
         }
-        this.bindings.bindings[RouteConstant.ROUTE_LOG_INPUT]?.let {
-            if (it.destination == null || it.destination == RouteConstant.ROUTE_LOG_INPUT) {
+        if (this.streamFunctionProperties.definition != null && this.streamFunctionProperties.definition.isNotBlank()) {
+            this.streamFunctionProperties.definition += ";${RouteConstant.ROUTE_LOG_INPUT}"
+        } else {
+            this.streamFunctionProperties.definition = RouteConstant.ROUTE_LOG_INPUT
+        }
+        this.bindingServiceProperties.bindings[routeLogConsumerBindName]?.let {
+            if (it.destination == null || it.destination == routeLogConsumerBindName) {
                 it.destination = RouteConstant.ROUTE_LOG_DESCRIPTION
             }
             it.contentType = MimeTypeUtils.APPLICATION_JSON_VALUE
@@ -45,11 +52,12 @@ constructor(private val bindings: BindingServiceProperties) {
         }
     }
 
-    @Bean
-    fun updateRouteConsumer(logAdapter: LogAdapter,
-                            objectMapper: ObjectMapper,
-                            logDomain: LogDomain,
-                            logServerCustomerConfiguration: LogServerCustomerConfiguration) =
-            RouteLogConsumer(logAdapter, objectMapper, logDomain, logServerCustomerConfiguration)
+    @Bean(RouteConstant.ROUTE_LOG_INPUT)
+    fun routeLogConsumer(
+        logAdapter: LogAdapter,
+        objectMapper: ObjectMapper,
+        logDomain: LogDomain,
+        logServerCustomerConfiguration: LogServerCustomerConfiguration
+    ) = RouteLogConsumer(logAdapter, objectMapper, logDomain, logServerCustomerConfiguration)
 
 }